+++ sw.js
@@ -1,11 +1,14 @@
 "use strict";
 
 // Bump this on any shell-file change so old installs pick up the update.
-var CACHE_NAME = "dd-shell-v14";
+var CACHE_NAME = "dd-shell-v15";
 
 var SHELL_FILES = [
   "./",
   "./index.html",
+  "./env.js",
+  "./items.js",
+  "./items-seed.json",
   "./style.css",
   "./app.js",
   "./theme.js",